fix(bench): cap max-parallel=5 + startup stagger + builder throttle-retry
Combat Bedrock account-wide TPM throttling in the per-PR agent-bench matrix.
Run 28967475174 failed 6/10 cells with "ModelError: Too many tokens" — all in
a 72s window when the uncapped 10-cell matrix hit InvokeModelWithResponseStream
at once. The builder step had NO retry (unlike the judge, which rode it out).

PART A — reduce & spread concurrent load (.github/workflows/agent-bench.yml):
  - Add strategy.max-parallel: 5 (keep fail-fast: false) so 10 cells run in two
    waves, halving peak concurrent token demand.
  - Add a DETERMINISTIC startup stagger before the builder's first Bedrock call,
    keyed to the 0-based matrix index (BENCH_CELL_INDEX = strategy.job-index).
    Each cell sleeps index * BENCH_STAGGER_STEP_SEC (default 7s, env-tunable) so
    a wave's opening calls fan out over ~0-28s instead of firing together.

PART B — throttle-retry resilience at every Bedrock call site:
  - New shared helper steps/lib/bedrock-retry.ts: the retry budget
    (INVOKE_MAX_ATTEMPTS=5, INVOKE_BACKOFF_MS=[5s,15s,40s,90s]), the
    throttle/transient classifier (isRetryableModelError), and a deep cause-chain
    error describer (describeModelError) — factored out of the judge.
  - Builder (2-agent-run.ts): add clientConfig {maxAttempts:8, retryMode:'adaptive'}
    and an app-level retry loop mirroring the judge (fresh agent per attempt so a
    mid-stream throw can't pollute the retry; module-scope token counters survive
    across attempts for the SIGTERM partial-envelope flush).
  - Judge (4-judge.ts): refactored to consume the shared helper (no behavior
    change) — single source of truth.
  - Analysis (lib/analysis.mjs, AWS-CLI converse): add AWS_RETRY_MODE=adaptive +
    AWS_MAX_ATTEMPTS=8 to the spawn (CLI analog of the SDK clientConfig) and widen
    the transient regex to catch the spaced "too many tokens/requests" message.

Scoring/reclassification of throttle-fails is intentionally NOT changed here.

Verified: tsc --noEmit clean; node --test steps/lib/*.test.mjs → 96/96 pass.
